refactor: rename fieldpath to hightlight api
