+++ connect/transforms/src/main/java/org/apache/kafka/connect/transforms/ExtractField.java
@@ -21,7 +21,7 @@ import org.apache.kafka.connect.connector.ConnectRecord;
 import org.apache.kafka.connect.data.Field;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.transforms.field.FieldPath;
+import org.apache.kafka.connect.transforms.field.SingleFieldPath;
 import org.apache.kafka.connect.transforms.field.FieldSyntaxVersion;
 import org.apache.kafka.connect.transforms.util.SimpleConfig;
 
@@ -50,12 +50,12 @@ public abstract class ExtractField<R extends ConnectRecord<R>> implements Transf
 
     private static final String PURPOSE = "field extraction";
 
-    private FieldPath fieldPath;
+    private SingleFieldPath fieldPath;
 
     @Override
     public void configure(Map<String, ?> props) {
         final SimpleConfig config = new SimpleConfig(CONFIG_DEF, props);
-        fieldPath = FieldPath.of(config.getString(FIELD_CONFIG), FieldSyntaxVersion.fromConfig(config));
+        fieldPath = SingleFieldPath.of(config.getString(FIELD_CONFIG), FieldSyntaxVersion.fromConfig(config));
     }
 
     @Override
