+++ connect/transforms/src/main/java/org/apache/kafka/connect/transforms/MaskField.java
@@ -23,7 +23,7 @@ import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
 import org.apache.kafka.connect.data.Values;
 import org.apache.kafka.connect.errors.DataException;
-import org.apache.kafka.connect.transforms.field.FieldPaths;
+import org.apache.kafka.connect.transforms.field.FieldPathGroup;
 import org.apache.kafka.connect.transforms.field.FieldSyntaxVersion;
 import org.apache.kafka.connect.transforms.util.NonEmptyListValidator;
 import org.apache.kafka.connect.transforms.util.SimpleConfig;
@@ -86,13 +86,13 @@ public abstract class MaskField<R extends ConnectRecord<R>> implements Transform
         REPLACEMENT_MAPPING_FUNC.put(BigInteger.class, BigInteger::new);
     }
 
-    private FieldPaths maskedFields;
+    private FieldPathGroup maskedFields;
     private String replacement;
 
     @Override
     public void configure(Map<String, ?> props) {
         final SimpleConfig config = new SimpleConfig(CONFIG_DEF, props);
-        maskedFields = FieldPaths.of(config.getList(FIELDS_CONFIG), FieldSyntaxVersion.fromConfig(config));
+        maskedFields = FieldPathGroup.of(config.getList(FIELDS_CONFIG), FieldSyntaxVersion.fromConfig(config));
         replacement = config.getString(REPLACEMENT_CONFIG);
     }
