+++ connect/transforms/src/main/java/org/apache/kafka/connect/transforms/field/MapValueUpdater.java
@@ -30,7 +30,7 @@ public interface MapValueUpdater {
     void apply(
             Map<String, Object> originalParent,
             Map<String, Object> updatedParent,
-            FieldPath fieldPath,
+            SingleFieldPath fieldPath,
             String fieldName
     );
 }
