+++ connect/transforms/src/main/java/org/apache/kafka/connect/transforms/field/StructValueUpdater.java
@@ -22,5 +22,5 @@ import org.apache.kafka.connect.data.Struct;
 @FunctionalInterface
 public interface StructValueUpdater {
 
-    void apply(Struct originalParent, Field originalField, Struct updatedParent, Field updatedField, FieldPath fieldPath);
+    void apply(Struct originalParent, Field originalField, Struct updatedParent, Field updatedField, SingleFieldPath fieldPath);
 }
